Python Exif Data Extractor
This Python Script Extract the photo Hidden Exif Data. Also it can extract the location of photo :: only if the location is tagged in photo while capturing the photo.

Note: location can be extracted if you embed the photo from social media sites, because when user upload the photo on Social Media then the location is been attached into the photo.